# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The Quack dashboard API.

    Every error status (400 and up) carries the error envelope, whatever the
    handler wrote. Unknown paths and wrong methods are 404s.

    The dashboard authenticates with the HttpOnly session cookie set by Discord
    sign-in (quack_session by default). Writes made with the cookie must come from
    an allowed Origin and echo the CSRF cookie (quack_csrf by default, also
    returned as csrf_token by /auth/me) in X-CSRF-Token. A bearer session token is
    accepted instead of the cookie and skips the CSRF check.

    Writes marked with an Idempotency-Key header are safe to retry with the same
    key: a completed request is replayed with Idempotency-Replayed: true, one
    still running is a 409 with Retry-After, and reusing a key for a different
    request is a 409.

    This file is generated from the Go route table. Do not edit it; run
    "go generate ./..." from apps/backend.
  title: Quack HTTP API
  version: "5"
paths:
...
  /guilds/{discordGuildID}/action-failures:
    get:
      operationId: listFailedActions
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - in: query
        name: limit
        schema:
          default: 50
          maximum: 100
          minimum: 1
          type: integer
      - in: query
        name: offset
        schema:
          default: 0
          maximum: 100000
          minimum: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailedActionListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: The queue of failed case actions
      tags:
      - action-failures
  /guilds/{discordGuildID}/action-failures/{executionID}/dismiss:
    post:
      operationId: dismissFailedAction
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - in: path
        name: executionID
        required: true
        schema:
          type: string
      - description: Makes the write safe to retry; reuse the key only for the identical
          request.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          description: Makes the write safe to retry; reuse the key only for the identical
            request.
          maxLength: 256
          type: string
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailedActionEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Remove a failure from the queue; its history stays
      tags:
      - action-failures
  /guilds/{discordGuildID}/action-failures/{executionID}/retry:
    post:
      operationId: retryFailedAction
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - in: path
        name: executionID
        required: true
        schema:
          type: string
      - description: Makes the write safe to retry; reuse the key only for the identical
          request.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          description: Makes the write safe to retry; reuse the key only for the identical
            request.
          maxLength: 256
          type: string
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailedActionEnvelope'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Requeue a failed action
      tags:
      - action-failures
...
  /guilds/{discordGuildID}/cases/{caseRef}/reversals:
    post:
      description: The body must set confirm.
      operationId: reverseCaseAction
      parameters:
      - in: path
        name: discordGuildID
        required: true
        schema:
          type: string
      - in: path
        name: caseRef
        required: true
        schema:
          type: string
      - description: Makes the write safe to retry; reuse the key only for the identical
          request.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          description: Makes the write safe to retry; reuse the key only for the identical
            request.
          maxLength: 256
          type: string
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiReverseActionRequest'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiActionEnvelope'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Too Many Requests
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Queue a timeout removal or unban
      tags:
      - cases
...
components:
  schemas:
    ApiActionEnvelope:
      properties:
        action:
          $ref: '#/components/schemas/QuackCaseActionExecution'
      required:
      - action
      type: object
...
    ApiErrorCode:
      enum:
      - validation_failed
      - authentication_required
      - reauthentication_required
      - authorization_denied
      - not_found
      - conflict
      - rate_limited
      - csrf_rejected
      - origin_rejected
      - body_too_large
      - dependency_unavailable
      - internal_error
      type: string
    ApiErrorDetail:
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        correlation_id:
          type: string
        message:
          type: string
        request_id:
          type: string
      required:
      - code
      - message
      - request_id
      - correlation_id
      type: object
    ApiErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDetail'
      required:
      - error
      type: object
    ApiFailedActionEnvelope:
      properties:
        action:
          $ref: '#/components/schemas/ApiFailedActionResponse'
      required:
      - action
      type: object
    ApiFailedActionListResponse:
      properties:
        executions:
          items:
            $ref: '#/components/schemas/ApiFailedActionResponse'
          type: array
        total:
          format: int64
          type: integer
      required:
      - executions
      - total
      type: object
    ApiFailedActionResponse:
      properties:
        action_type:
          $ref: '#/components/schemas/QuackActionType'
        attempt_count:
          minimum: 0
          type: integer
        case_id:
          type: string
        created_at:
          format: date-time
          type: string
        id:
          type: string
        last_error:
          type: string
        last_error_code:
          type: string
        max_retries:
          minimum: 0
          type: integer
        safe_for_retry:
          type: boolean
        status:
          $ref: '#/components/schemas/QuackActionExecutionStatus'
        updated_at:
          format: date-time
          type: string
      required:
      - id
      - case_id
      - action_type
      - status
      - attempt_count
      - max_retries
      - safe_for_retry
      - created_at
      - updated_at
      type: object
...
    ApiReverseActionRequest:
      properties:
        action_type:
          $ref: '#/components/schemas/QuackActionType'
        appeal_id:
          type:
          - "null"
          - string
        confirm:
          type: boolean
        original_execution_id:
          type: string
      type: object
...
    QuackActionExecutionStatus:
      enum:
      - pending
      - running
      - succeeded
      - failed
      - retrying
      - cancelled
      type: string
    QuackActionType:
      enum:
      - send_dm
      - timeout_user
      - kick_user
      - ban_user
      - remove_timeout
      - unban_user
      type: string
...
    QuackCaseActionExecution:
      properties:
        ActionType:
          $ref: '#/components/schemas/QuackActionType'
        AttemptCount:
          minimum: 0
          type: integer
        CaseID:
          type: string
        ConfigSnapshotJSON:
          type: string
        CorrelationID:
          type: string
        CreatedAt:
          format: date-time
          type: string
        DismissedAt:
          format: date-time
          type:
          - "null"
          - string
        DismissedByDiscordUserID:
          type: string
        FinishedAt:
          format: date-time
          type:
          - "null"
          - string
        ID:
          type: string
        IdempotencyKey:
          type: string
        LastError:
          type: string
        LastErrorCode:
          type: string
        LeaseExpiresAt:
          format: date-time
          type:
          - "null"
          - string
        LeaseToken:
          type: string
        MaxRetries:
          minimum: 0
          type: integer
        NextRetryAt:
          format: date-time
          type:
          - "null"
          - string
        NotificationType:
          type: string
        NotifyUser:
          type: boolean
        Position:
          type: integer
        RetryBackoffMS:
          type: integer
        ReversalAppealID:
          type:
          - "null"
          - string
        ReversalOfExecutionID:
          type:
          - "null"
          - string
        SafeForRetry:
          type: boolean
        StartedAt:
          format: date-time
          type:
          - "null"
          - string
        Status:
          $ref: '#/components/schemas/QuackActionExecutionStatus'
        TemplateActionID:
          type:
          - "null"
          - string
        UpdatedAt:
          format: date-time
          type: string
      type: object
...
  securitySchemes:
    metricsKey:
      description: The metrics key (api.metrics_token).
      in: header
      name: X-Quack-Metrics-Key
      type: apiKey
    opsKey:
      description: The operator key (api.ops_token).
      in: header
      name: X-Quack-Ops-Key
      type: apiKey
    sessionBearer:
      description: The session ID as a bearer token, for callers that are not browsers.
      scheme: bearer
      type: http
    sessionCookie:
      description: The session cookie set by Discord sign-in. Its name is configurable
        (auth.session_cookie_name).
      in: cookie
      name: quack_session
      type: apiKey